requirements.txt
PyVTK
MeshPy-2016.1
NumPy
trimesh
bidict
tqdm